Move screens to its own folder in ui package
